tests: Use g_assert_true()
Replace g_assert() with g_assert_true().  g_assert() is a no-op when
compiled with G_DISABLE_ASSERT, so it shouldn't be used in unit tests.

As of glib-2.58, g_test_init() will print an error and exit if compiled
with G_DISABLE_ASSERT.  So it is not longer an issue with new glib.
But better be safe and write tests the right way.