project: Add Authors list to license blurbs
Many firms keep the copyright of the code written by their employees.
So it's better to keep a list of authors for reference.

You may find 'git blame' interesting if you need accurate details
of who wrote a specific line of code (Please note: As many developers
make use of this data, it is recommended to not make whiteline
changes as much as possible).