build: Move version information to version.h
version information is regenerated on every commit.  This marks
config.h as modified on every commit, and as config.h is included
in every source file, a full rebuild is intiated.

So move version information generated to version.h and include the
file only for the files that requires it.